[RAS] Refactor cost model
