+++ gluten-core/src/main/scala/org/apache/gluten/planner/cost/GlutenCostModel.scala
@@ -17,20 +17,16 @@
 package org.apache.gluten.planner.cost
 
 import org.apache.gluten.GlutenConfig
-import org.apache.gluten.extension.columnar.enumerated.RemoveFilter
-import org.apache.gluten.extension.columnar.transition.{ColumnarToRowLike, RowToColumnarLike}
-import org.apache.gluten.planner.plan.GlutenPlanModel.GroupLeafExec
-import org.apache.gluten.ras.{Cost, CostModel}
-import org.apache.gluten.utils.PlanUtil
+import org.apache.gluten.ras.CostModel
 
 import org.apache.spark.internal.Logging
-import org.apache.spark.sql.execution.{ColumnarToRowExec, RowToColumnarExec, SparkPlan}
+import org.apache.spark.sql.execution.SparkPlan
 import org.apache.spark.sql.utils.ReflectionUtil
 
 object GlutenCostModel extends Logging {
   def find(): CostModel[SparkPlan] = {
-    val aliases: Map[String, Class[_ <: CostModel[SparkPlan]]] = Map(
-      "rough" -> classOf[RoughCostModel])
+    val aliases: Map[String, Class[_ <: CostModel[SparkPlan]]] =
+      Map("legacy" -> classOf[LegacyCostModel], "rough" -> classOf[RoughCostModel])
     val aliasOrClass = GlutenConfig.getConf.rasCostModel
     val clazz: Class[_ <: CostModel[SparkPlan]] = if (aliases.contains(aliasOrClass)) {
       aliases(aliasOrClass)
@@ -45,55 +41,5 @@ object GlutenCostModel extends Logging {
     model
   }
 
-  def rough(): CostModel[SparkPlan] = new RoughCostModel()
-
-  private class RoughCostModel extends CostModel[SparkPlan] {
-    private val infLongCost = Long.MaxValue
-
-    override def costOf(node: SparkPlan): GlutenCost = node match {
-      case _: GroupLeafExec => throw new IllegalStateException()
-      case _ => GlutenCost(longCostOf(node))
-    }
-
-    private def longCostOf(node: SparkPlan): Long = node match {
-      case n =>
-        val selfCost = selfLongCostOf(n)
-
-        // Sum with ceil to avoid overflow.
-        def safeSum(a: Long, b: Long): Long = {
-          assert(a >= 0)
-          assert(b >= 0)
-          val sum = a + b
-          if (sum < a || sum < b) Long.MaxValue else sum
-        }
-
-        (n.children.map(longCostOf).toList :+ selfCost).reduce(safeSum)
-    }
-
-    // A very rough estimation as of now. The cost model basically considers any
-    // fallen back ops as having extreme high cost so offloads computations as
-    // much as possible.
-    private def selfLongCostOf(node: SparkPlan): Long = {
-      node match {
-        case _: RemoveFilter.NoopFilter =>
-          // To make planner choose the tree that has applied rule PushFilterToScan.
-          0L
-        case ColumnarToRowExec(child) => 10L
-        case RowToColumnarExec(child) => 10L
-        case ColumnarToRowLike(child) => 10L
-        case RowToColumnarLike(child) => 10L
-        case p if PlanUtil.isGlutenColumnarOp(p) => 10L
-        case p if PlanUtil.isVanillaColumnarOp(p) => 1000L
-        // Other row ops. Usually a vanilla row op.
-        case _ => 1000L
-      }
-    }
-
-    override def costComparator(): Ordering[Cost] = Ordering.Long.on {
-      case GlutenCost(value) => value
-      case _ => throw new IllegalStateException("Unexpected cost type")
-    }
-
-    override def makeInfCost(): Cost = GlutenCost(infLongCost)
-  }
+  def legacy(): CostModel[SparkPlan] = new LegacyCostModel()
 }
